refactor: better classes
